fixed gitignore for runs/ dir
