+++ Archive/Train.py
@@ -1,9 +1,9 @@
 import os
 import matplotlib.pyplot as plt
 import numpy as np
-from Loss import compute_loss
-from YOLOv1 import YOLOv1
-from Dataset import CUB200
+from Archive.Loss import compute_loss
+from Archive.YOLOv1 import YOLOv1
+from Archive.Dataset import CUB200
 import torch
 from torch.utils.data import DataLoader
