+++ cub200_yolo/cub200.yaml
@@ -1,0 +1,208 @@
+train: ./images/train
+val: ./images/val
+
+# Number of classes for detection
+nc: 200
+
+# Class names for detection
+names:
+  - Black footed Albatross
+  - Laysan Albatross
+  - Sooty Albatross
+  - Groove billed Ani
+  - Crested Auklet
+  - Least Auklet
+  - Parakeet Auklet
+  - Rhinoceros Auklet
+  - Brewer Blackbird
+  - Red winged Blackbird
+  - Rusty Blackbird
+  - Yellow headed Blackbird
+  - Bobolink
+  - Indigo Bunting
+  - Lazuli Bunting
+  - Painted Bunting
+  - Cardinal
+  - Spotted Catbird
+  - Gray Catbird
+  - Yellow breasted Chat
+  - Eastern Towhee
+  - Chuck will Widow
+  - Brandt Cormorant
+  - Red faced Cormorant
+  - Pelagic Cormorant
+  - Bronzed Cowbird
+  - Shiny Cowbird
+  - Brown Creeper
+  - American Crow
+  - Fish Crow
+  - Black billed Cuckoo
+  - Mangrove Cuckoo
+  - Yellow billed Cuckoo
+  - Gray crowned Rosy Finch
+  - Purple Finch
+  - Northern Flicker
+  - Acadian Flycatcher
+  - Great Crested Flycatcher
+  - Least Flycatcher
+  - Olive sided Flycatcher
+  - Scissor tailed Flycatcher
+  - Vermilion Flycatcher
+  - Yellow bellied Flycatcher
+  - Frigatebird
+  - Northern Fulmar
+  - Gadwall
+  - American Goldfinch
+  - European Goldfinch
+  - Boat tailed Grackle
+  - Eared Grebe
+  - Horned Grebe
+  - Pied billed Grebe
+  - Western Grebe
+  - Blue Grosbeak
+  - Evening Grosbeak
+  - Pine Grosbeak
+  - Rose breasted Grosbeak
+  - Pigeon Guillemot
+  - California Gull
+  - Glaucous winged Gull
+  - Heermann Gull
+  - Herring Gull
+  - Ivory Gull
+  - Ring billed Gull
+  - Slaty backed Gull
+  - Western Gull
+  - Anna Hummingbird
+  - Ruby throated Hummingbird
+  - Rufous Hummingbird
+  - Green Violetear
+  - Long tailed Jaeger
+  - Pomarine Jaeger
+  - Blue Jay
+  - Florida Jay
+  - Green Jay
+  - Dark eyed Junco
+  - Tropical Kingbird
+  - Gray Kingbird
+  - Belted Kingfisher
+  - Green Kingfisher
+  - Pied Kingfisher
+  - Ringed Kingfisher
+  - White breasted Kingfisher
+  - Red legged Kittiwake
+  - Horned Lark
+  - Pacific Loon
+  - Mallard
+  - Western Meadowlark
+  - Hooded Merganser
+  - Red breasted Merganser
+  - Mockingbird
+  - Nighthawk
+  - Clark Nutcracker
+  - White breasted Nuthatch
+  - Baltimore Oriole
+  - Hooded Oriole
+  - Orchard Oriole
+  - Scott Oriole
+  - Ovenbird
+  - Brown Pelican
+  - White Pelican
+  - Western Wood Pewee
+  - Sayornis
+  - American Pipit
+  - Whip poor Will
+  - Horned Puffin
+  - Common Raven
+  - White necked Raven
+  - American Redstart
+  - Geococcyx
+  - Loggerhead Shrike
+  - Great Grey Shrike
+  - Baird Sparrow
+  - Black throated Sparrow
+  - Brewer Sparrow
+  - Chipping Sparrow
+  - Clay colored Sparrow
+  - House Sparrow
+  - Field Sparrow
+  - Fox Sparrow
+  - Grasshopper Sparrow
+  - Harris Sparrow
+  - Henslow Sparrow
+  - Le Conte Sparrow
+  - Lincoln Sparrow
+  - Nelson Sharp tailed Sparrow
+  - Savannah Sparrow
+  - Seaside Sparrow
+  - Song Sparrow
+  - Tree Sparrow
+  - Vesper Sparrow
+  - White crowned Sparrow
+  - White throated Sparrow
+  - Cape Glossy Starling
+  - Bank Swallow
+  - Barn Swallow
+  - Cliff Swallow
+  - Tree Swallow
+  - Scarlet Tanager
+  - Summer Tanager
+  - Artic Tern
+  - Black Tern
+  - Caspian Tern
+  - Common Tern
+  - Elegant Tern
+  - Forsters Tern
+  - Least Tern
+  - Green tailed Towhee
+  - Brown Thrasher
+  - Sage Thrasher
+  - Black capped Vireo
+  - Blue headed Vireo
+  - Philadelphia Vireo
+  - Red eyed Vireo
+  - Warbling Vireo
+  - White eyed Vireo
+  - Yellow throated Vireo
+  - Bay breasted Warbler
+  - Black and white Warbler
+  - Black throated Blue Warbler
+  - Blue winged Warbler
+  - Canada Warbler
+  - Cape May Warbler
+  - Cerulean Warbler
+  - Chestnut sided Warbler
+  - Golden winged Warbler
+  - Hooded Warbler
+  - Kentucky Warbler
+  - Magnolia Warbler
+  - Mourning Warbler
+  - Myrtle Warbler
+  - Nashville Warbler
+  - Orange crowned Warbler
+  - Palm Warbler
+  - Pine Warbler
+  - Prairie Warbler
+  - Prothonotary Warbler
+  - Swainson Warbler
+  - Tennessee Warbler
+  - Wilson Warbler
+  - Worm eating Warbler
+  - Yellow Warbler
+  - Northern Waterthrush
+  - Louisiana Waterthrush
+  - Bohemian Waxwing
+  - Cedar Waxwing
+  - American Three toed Woodpecker
+  - Pileated Woodpecker
+  - Red bellied Woodpecker
+  - Red cockaded Woodpecker
+  - Red headed Woodpecker
+  - Downy Woodpecker
+  - Bewick Wren
+  - Cactus Wren
+  - Carolina Wren
+  - House Wren
+  - Marsh Wren
+  - Rock Wren
+  - Winter Wren
+  - Common Yellowthroat
